Add install script for doc, make bin install more intelligent.

+++ release/scripts/bin-install.sh
@@ -5,6 +5,8 @@ if [ "`id -u`" != "0" ]; then
 	exit 1
 fi
 echo "You are about to extract the base distribution into / - are you SURE"
-echo "you want to do this over your installed system?  If not, hit ^C now!"
-read junk
-cat bin.?? | tar --unlink -xpzf - -C /
+echo -n "you want to do this over your installed system (y/n)? "
+read ans 
+if [ "$ans" = "y" ]; then
+	cat bin.?? | tar --unlink -xpzf - -C /
+fi
